test for computeScore function
